requirements.txt
torch
torchvision
flwr
numpy
pandas
opencv-python
pydicom
tqdm
scikit-learn
Pillow
psutil 